requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
spotipy
python-dotenv
textblob
transformers
torch
numpy
scikit-learn
librosa
soundfile
scipy
midiutil

